# DOCS/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/v1/user/leaves": {
      "get": {
        "tags": [
          "leave request"
        ],
        "summary": "Get user's leave requests",
        "description": "Retrieve leave requests submitted by the user.",
        "operationId": "getUserLeaveRequests",
        "security": [
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response. Retrieves user's leave requests.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "Your account is not exist."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "OTP is invalid."
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "leave request"
        ],
        "summary": "Request Leave",
        "description": "Request leave with specified start and end dates.",
        "operationId": "requestLeave",
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "description": "Object containing details for leave request.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-06-15T13:17:28.475Z"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-06-15T13:17:28.475Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leave request successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string",
                      "example": "All data successfully retrieved."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "requestDate": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-06-15T13:17:28.475Z"
                        },
                        "userId": {
                          "type": "integer",
                          "example": 1
                        },
                        "startDate": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true,
                          "example": null
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true,
                          "example": null
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-06-15T13:17:28.475Z"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-06-15T13:17:28.475Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "Your account is not exist."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "OTP is invalid."
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FAIL": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "FAIL"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}